feature: rename resource form component
